+++ Editor/Package.cs
@@ -1,10 +1,11 @@
 ﻿
 using UnityEditor;
 
-namespace Hananoki.CustomProjectBrowser {
+namespace HananokiEditor.CustomProjectBrowser {
   public static class Package {
     public const string name = "CustomProjectBrowser";
+    public const string nameNicify = "Custom Project Browser";
     public const string editorPrefName = "Hananoki.CustomProjectBrowser";
-    public const string version = "0.5.13-preview";
+    public const string version = "0.5.14-preview";
   }
 }